requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
engineering-notation==0.10.0
idna==3.10
mysql-connector-python==9.1.0
Pillow==9.5.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
tk==0.1.0
tk-tools==0.16.0
urllib3==2.2.3
